Replace TodoMVC with basic Hello World example
Since it's a pain to remove all TodoMVC files and references when starting a new project, I'm moving it to a separate repo. The react-starter repo now consists of a basic Hello World app.

I'm also removing a couple of modules and files that are not used here anymore, such as:

- babel-polyfill
- classnames
- isomorphic-fetch
- mock API
- nock
- redux-mock-store
- redux-saga
- request util

I will try to keep this project as slim as possible, but consisting of the things I use in all my React projects. For a more complete implementation of Redux, Redux Saga, async actions etc, see my TodoMVC example.